refactor: element define
- abstract setupTokenHighlights
- run setupTokenHighlights in element define

BREAKING CHANGE:

`config.extendTokenTypes` got renamed to `config.languageTokens`
