+++ docs/index.html
@@ -23,7 +23,7 @@
       window.SyntaxHighlightElement = window.SyntaxHighlightElement || {};
       window.SyntaxHighlightElement.config = {
         // Optional: language specific token type overwrites
-        extendTokenTypes: {
+        languageTokens: {
           css: ['important'],
         }
       }
