+++ index.html
@@ -33,7 +33,7 @@
       window.SyntaxHighlightElement = window.SyntaxHighlightElement || {};
       window.SyntaxHighlightElement.config = {
         // Optional: language specific token type overwrites
-        extendTokenTypes: {
+        languageTokens: {
           css: ['important'],
         }
       }
